filters, sensors, fuzzy_controllers, loop integration

+++ particle.h
@@ -2,9 +2,9 @@
 #define __PARTICLE_H__
 
 typedef struct particle {
-  int x;
-  int y;
-  int theta;
+  double x;
+  double y;
+  double theta;
 } particle;
 
 #endif
